Add reminders extension. Closes #17
The reminders extension allows users to configure one-time or
repeating reminders in text and voice channels. In text channels, the
message will be send after a specified delay, and in voice channels, the
message will be read as TTS. Repeating reminders are saved to a database
for persistence across bot invocations.

The version of discord.py had to be bumped to support some new command
type annotation features.