Fix TTS not always working. Closes #46
The issue was writing the MP3 to file and playing the sound from the
file path happened fast enough that the file buffer was not always
flushed resulting in ffmpeg reading a corrupted file. Fix was to simply
flush the file before returning the file path.

Also changed /tts to /speak because Discord has a /tts command already.