Add random offset for repeating reminders (#50)
The goal is to minimize collisions between reminders that are multiples
of each others period. This is a limitation of the reminders system not
having a queue which particularly matters for voice reminders because
playing a sound will cancel and currently playing sounds.

For example, a reminder with a period of 1 hour and another with a
period of 2 hours will have collisions where they trigger at the same
time. If these are voice reminders, one will cancel the other. This is
particularly a problem for repeating reminders, because when the bot
starts up, it starts all repeating reminders at more or less the same
time.

The fix is to add a random sleep in the interval [1, reminder delay]
before starting the first loop of the task for that reminder. This is
only done for repeating reminders.