Fix coverage failure with birthday task (#87)
I refactored the code so there is no longer a branch that can be
uncovered. The last failure I see occured when the test ran at 00:00:22
GMT so I suspect the failure has to do with the
now.minute > BIRTHDAY_CHECK_MINUTE as both were 0 in the failed CI run.
I changed this check to be >=. It was original just > to give 60 seconds
of wiggle room on when the tasks starts if the bot happened to be
started at the BIRTHDAY_CHECK time.