Add coverage pragmas for tests with nested contexts
I think this may be a regression in coverage. It seems there are cases
where a nested context manager is treated as a branch where the branch
is uncovered. I don't have time at the moment to look into this, so just
fixing with pragmas. (I will look into it separately and open a
coverage issue.)